Decouple room watcher and socket events
Previously, the server would listen to a specific message to be emitted
by a client upon them entering a room. The server would then start
watching all database rooms for changes and would have to do a check to
ignore changes in other rooms. This was undesirable for two reasons:

 - This listener would activate on every new client in the room,
   emitting as many new messages as there were clients, whereas any
   message should only be emitted once.
 - Because the listener listened to all changes in any room, the load of
   this function grew exponentially with the number of clients
   connected.

Now, the server listens globally for any changes in a room, then emits
messages directly to that room, reducing code complexity. As a bonus,
the client does not need to emit any message to activate the server's
emitting of messages.